# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AI Business Platform API",
        "description": "Authentication & RBAC endpoints for the AI Business Platform. See ARCHITECTURE.md, API.md, and BACKEND.md in the repository root for the full system design.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current server"
        }
    ],
    "paths": {
...
        "/api/v1/employees/{employee}/notes": {
            "get": {
                "tags": [
                    "Employees / Notes"
                ],
                "summary": "List notes for an employee",
                "operationId": "5a52e642b93b530360352d2d935b0522",
                "parameters": [
                    {
                        "name": "employee",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notes returned"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Employees / Notes"
                ],
                "summary": "Add a note to an employee",
                "operationId": "0d88a61145808c385bc59cf4cbed46e4",
                "parameters": [
                    {
                        "name": "employee",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "note"
                                ],
                                "properties": {
                                    "note": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Note created"
                    },
                    "403": {
                        "description": "Not authorized to add notes for this employee"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/positions": {
            "get": {
                "tags": [
                    "Employees / Positions"
                ],
                "summary": "List positions",
                "operationId": "951f16c30e98098d2bebbbbfda1175df",
                "responses": {
                    "200": {
                        "description": "Positions returned"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Employees / Positions"
                ],
                "summary": "Create a position",
                "operationId": "f924e0831a9362a60a2a8e2e5da0bc1b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Software Engineer"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Position created"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/positions/{position}": {
            "get": {
                "tags": [
                    "Employees / Positions"
                ],
                "summary": "Get a single position",
                "operationId": "cec089064500c3ae43203fecd497dbfa",
                "parameters": [
                    {
                        "name": "position",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Position returned"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Employees / Positions"
                ],
                "summary": "Delete a position",
                "operationId": "35bcb440b1a754f6da7b4e9b1843d710",
                "parameters": [
                    {
                        "name": "position",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Position deleted"
                    },
                    "403": {
                        "description": "Missing positions.manage permission"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Employees / Positions"
                ],
                "summary": "Update a position",
                "operationId": "bd45f1de598db8fa0b310bb004790cba",
                "parameters": [
                    {
                        "name": "position",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Position updated"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Send the token returned by /api/v1/auth/register or /api/v1/auth/login as 'Authorization: Bearer {token}'.",
                "bearerFormat": "Sanctum personal access token",
                "scheme": "bearer"
            }
        }
    },
...
}